added token

+++ contracts/Token.sol
@@ -3,11 +3,11 @@ pragma solidity ^0.8.4;
 
 library Balances {
     function move(
-        mapping(address => uint256) storage balances, 
-        address from, address to, uint amount
-        ) 
-        internal 
-        {
+        mapping(address => uint256) storage balances,
+        address from,
+        address to,
+        uint256 amount
+    ) internal {
         require(balances[from] >= amount);
         require(balances[to] + amount >= balances[to]);
         balances[from] -= amount;
@@ -18,18 +18,25 @@ library Balances {
 contract Token {
     mapping(address => uint256) balances;
     using Balances for *;
-    mapping(address => mapping (address => uint256)) allowed;
+    mapping(address => mapping(address => uint256)) allowed;
 
-    event Transfer(address from, address to, uint amount);
-    event Approval(address owner, address spender, uint amount);
+    event Transfer(address from, address to, uint256 amount);
+    event Approval(address owner, address spender, uint256 amount);
 
-    function transfer(address to, uint amount) public returns (bool success) {
+    function transfer(address to, uint256 amount)
+        public
+        returns (bool success)
+    {
         balances.move(msg.sender, to, amount);
         emit Transfer(msg.sender, to, amount);
         return true;
     }
 
-    function transferFrom(address from, address to, uint amount) public returns (bool success) {
+    function transferFrom(
+        address from,
+        address to,
+        uint256 amount
+    ) public returns (bool success) {
         require(allowed[from][msg.sender] >= amount);
         allowed[from][msg.sender] -= amount;
         balances.move(from, to, amount);
@@ -37,14 +44,21 @@ contract Token {
         return true;
     }
 
-    function approve(address spender, uint tokens) public returns (bool success) {
+    function approve(address spender, uint256 tokens)
+        public
+        returns (bool success)
+    {
         require(allowed[msg.sender][spender] == 0, "");
         allowed[msg.sender][spender] = tokens;
         emit Approval(msg.sender, spender, tokens);
         return true;
     }
 
-    function balanceOf(address tokenOwner) public view returns (uint balance) {
+    function balanceOf(address tokenOwner)
+        public
+        view
+        returns (uint256 balance)
+    {
         return balances[tokenOwner];
     }
 }
